[Changed] reorganized entity binding to another file
