+++ engine/bindings/private/entity/entity_bind.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include "wren_common.hpp"
+
+void BindEntityAPI(wren::ForeignModule& module);
